- Support part of Xcode 7.1 beta. - Support Apple TV launch image. (Launch image only, not support icons) - Support new CarPlay Icons (60x60x2, 60x60x3) - Bug fix. - This application will not support "Brand Assets".

+++ ios-image-util/src/org/gootara/ios/image/util/ui/SplitterFrame.java
@@ -674,11 +674,13 @@ public class SplitterFrame extends JFrame {
 							return false;
 						}
 					});
-					for (String filename : filenames) {
-						File f = new File(targetDirectory, filename);
-						if (f.exists() && f.isFile()) {
-							if (!f.delete()) {
-								System.err.println("Could not delete file - " + f.getAbsolutePath());
+					if (filenames != null) {
+						for (String filename : filenames) {
+							File f = new File(targetDirectory, filename);
+							if (f.exists() && f.isFile()) {
+								if (!f.delete()) {
+									System.err.println("Could not delete file - " + f.getAbsolutePath());
+								}
 							}
 						}
 					}
